try with void const function but int *array is not const

+++ anikanov.vladislav/P4/check.cpp
@@ -5,7 +5,7 @@
 bool isNumber(const char *str)
 {
   char c = 0;
-  for (int i = 0; i < strlen(str); ++i) {
+  for (size_t i = 0; i < strlen(str); ++i) {
     c = str[i];
     if (!std::isdigit(c)) {
       return false;
